requirements.txt
airbrake==2.1.0
amqp==2.5.0
anyjson==0.3.3
asgiref==2.3.2
async-timeout==3.0.1
attrs==19.1.0
autobahn==19.2.1
Automat==0.7.0
backcall==0.1.0
billiard==3.5.0.5
bleach==3.1.0
cachetools==3.1.0
celery==4.2.0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
colorama==0.4.1
constantly==15.1.0
dash==0.39.0
dash-core-components==0.44.0
dash-html-components==0.14.0
dash-renderer==0.20.0
dash-table==3.6.0
decorator==4.3.2
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.1.4
django-appconf==1.0.2
django-bootstrap-breadcrumbs==0.9.1
django-bootstrap4==0.0.8
django-chartjs==1.3
django-cors-headers==2.4.0
django-crispy-forms==1.7.2
django-export-xls==0.1.1
django-extensions==2.1.7
django-filter==2.1.0
django-formtools==2.1
django-google-charts==0.1.1
django-googledrive-storage==1.3.5
django-heroku==0.3.1
django-ordered-model==3.1.1
django-pandas==0.6.0
django-plotly-dash==0.9.8
django-redis-cache==2.0.0
django-tables2==2.0.5
django-template-maths==0.2.0
django-user-accounts==2.1.0
django-widget-tweaks==1.4.3
dpd-components==0.1.0
entrypoints==0.3
Flask==1.0.2
Flask-Compress==1.4.0
google-api-python-client==1.7.8
google-auth==1.6.3
google-auth-httplib2==0.0.3
gunicorn==19.9.0
httplib2==0.12.1
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
ipykernel==5.1.1
ipython==7.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
itsdangerous==1.1.0
jedi==0.13.3
Jinja2==2.10
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
kombu==4.6.3
libmagic==1.0
MarkupSafe==1.1.1
mistune==0.8.4
nbconvert==5.5.0
nbformat==4.4.0
notebook==5.7.8
numpy==1.16.2
oauth2client==4.1.3
pandas==0.24.1
pandocfilters==1.4.2
parso==0.4.0
pickleshare==0.7.5
plotly==3.6.1
prometheus-client==0.7.0
prompt-toolkit==2.0.9
psycopg2==2.7.6.1
pyasn1==0.4.5
pyasn1-modules==0.2.4
Pygments==2.4.2
PyHamcrest==1.9.0
pyrsistent==0.14.11
python-dateutil==2.8.0
python-magic==0.4.15
pytz==2018.7
pywinpty==0.5.5
pyzmq==18.0.1
qtconsole==4.5.1
redis==3.2.1
requests==2.21.0
retrying==1.3.3
rsa==4.0
Send2Trash==1.5.0
six==1.12.0
terminado==0.8.2
testpath==0.4.2
tornado==6.0.2
traitlets==4.3.2
txaio==18.8.1
uritemplate==3.0.0
urllib3==1.24.1
vine==1.3.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
whitenoise==4.1.2
widgetsnbextension==3.4.2
xlrd==1.2.0
xlwt==1.3.0
zope.interface==4.6.0
